moved flyzone code into seperate file, minor refactor and bug fixes
