+++ src/graph/mod.rs
@@ -5,10 +5,12 @@ use super::*;
 mod test;
 
 mod connection;
+mod flyzones;
 mod node;
 mod point;
-pub mod util;
 mod vertex;
+
+pub mod util;
 
 pub use graph::util::*;
 use obj::{Location, Obstacle};
@@ -151,85 +153,12 @@ impl Pathfinder {
         self.nodes.clear();
         self.find_origin();
         for i in 0..self.obstacles.len() {
-            let mut node = Node::from_obstacle(&self.obstacles[i], &self.origin);
+            let mut node = Node::from_obstacle(&self.obstacles[i], &self.origin, self.buffer);
             self.insert_flyzone_sentinel(&mut node);
             self.nodes.push(Rc::new(RefCell::new(node)));
         }
         for i in 0..self.flyzones.len() {
              self.virtualize_flyzone(i);
-        }
-    }
-
-    pub fn virtualize_flyzone(&mut self, index: usize) {
-        let flyzone = &self.flyzones[index];
-        // convert flyzone to points
-        let mut flyzone_points = Vec::new();
-        for location in flyzone {
-            let point = Point::from_location(&location, &self.origin);
-            flyzone_points.push(point);
-        }
-        // determine flyzone directions
-        let size = flyzone.len() as isize - 1;
-        let (clockwise, _) = vertex_direction(&flyzone_points);
-        let (direction, mut iter): (isize, isize) = if clockwise == true {
-            (1, 0)
-        } else {
-            (-1, size)
-        };
-        // edge conditions for flyzone
-        while (iter <= size) && (iter > -1) {
-            let (prev, next) = if iter == 0 {
-                (size, iter + 1)
-            } else if iter == size {
-                (iter - 1, 0)
-            } else {
-                (iter - 1, iter + 1)
-            };
-            // initalize obstacle location
-            let a = flyzone_points[prev as usize];
-            let vertex = flyzone_points[iter as usize];
-            let b = flyzone_points[next as usize];
-            let vec_a = (a.x - vertex.x, a.y - vertex.y);
-            let vec_b = (b.x - vertex.x, b.y - vertex.y);
-            let mag_a = ((vec_a.0).powi(2) + (vec_a.1).powi(2)).sqrt();
-            let mag_b = ((vec_b.0).powi(2) + (vec_b.1).powi(2)).sqrt();
-            let bisect = (
-                mag_b * vec_a.0 + mag_a * vec_b.0,
-                mag_b * vec_a.1 + mag_a * vec_b.1,
-            );
-            let mag_bisection = ((bisect.0).powi(2) + (bisect.1).powi(2)).powf(0.5);
-            let bisection = (bisect.0 / mag_bisection, bisect.1 / mag_bisection);
-            let theta = ((vec_a.0 * vec_b.0 + vec_a.1 * vec_b.1) / (mag_a * mag_b)).acos();
-            // section direction
-            let (iter_clockwise, straight) = vertex_direction(&vec![a, vertex, b]);
-            // straight line condition
-            if straight == true {
-                println!("Straight!");
-            } else {
-                let d = if (iter_clockwise == false && direction == 1)
-                    || (iter_clockwise == true && direction == -1)
-                {
-                    TURNING_RADIUS
-                } else {
-                    TURNING_RADIUS / ((theta / 2f32).sin())
-                };
-
-                if d > mag_a || d > mag_b {
-                    println!("small angle");
-                } else {
-                    // normal angle node
-                    let dis = d;
-                    let center = Point::new(
-                        dis * bisection.0 + vertex.x,
-                        dis * bisection.1 + vertex.y,
-                        0f32,
-                    );
-                    //println!("center: {:?}", center);
-                    let virt_ob = Node::new(center, TURNING_RADIUS, 0f32);
-                    self.nodes.push(Rc::new(RefCell::new(virt_ob)));
-                }
-            }
-            iter += direction;
         }
     }
 
@@ -271,73 +200,6 @@ impl Pathfinder {
             self.origin.lat_degree(),
             self.origin.lon_degree()
         );
-    }
-
-    // determines vertices of node and flyzone intersection
-    fn insert_flyzone_sentinel(&mut self, node: &mut Node) {
-        let center: Point = node.origin;
-        let r: f32 = node.radius;
-        for flyzone in self.flyzones.iter() {
-            let size = flyzone.len();
-            // iterate node over all vertices
-            for i in 0..size {
-                let mut v1 = flyzone[i];
-                let mut v2 = flyzone[(i + 1) % size];
-                let (x, y, dist_squared, end) = intersect_distance(
-                    &Point::from_location(&v1, &self.origin),
-                    &Point::from_location(&v2, &self.origin),
-                    &center,
-                );
-                let dist = dist_squared.sqrt();
-                // println!("dist: {:?}",dist);
-                // check intersect is true
-                if dist > r {
-                    continue;
-                }
-                // determine both intersect angles in left and right ring
-                let theta = (dist / r).acos(); //check
-                let dx = x - center.x; //check
-                let dy = y - center.y; //check
-                let phi = dy.atan2(dx); //check
-                let a = phi + theta;
-                let b = phi - theta;
-                let vertex_a = Rc::new(RefCell::new(Vertex::new_sentinel(
-                    &mut self.num_vertices,
-                    node,
-                    a,
-                )));
-                let vertex_b = Rc::new(RefCell::new(Vertex::new_sentinel(
-                    &mut self.num_vertices,
-                    node,
-                    b,
-                )));
-
-                let a_lat = vertex_a
-                    .borrow()
-                    .location
-                    .to_location(&self.origin)
-                    .lat_degree();
-                let a_lon = vertex_a
-                    .borrow()
-                    .location
-                    .to_location(&self.origin)
-                    .lon_degree();
-                let b_lat = vertex_b
-                    .borrow()
-                    .location
-                    .to_location(&self.origin)
-                    .lat_degree();
-                let b_lon = vertex_b
-                    .borrow()
-                    .location
-                    .to_location(&self.origin)
-                    .lon_degree();
-                println!("flyzone/node vertices: {:?},{:?}", a_lat, a_lon);
-                println!("flyzone/node vertices: {:?},{:?}", b_lat, b_lon);
-                node.insert_vertex(vertex_a);
-                node.insert_vertex(vertex_b);
-            }
-        }
     }
 
     // Generate all valid possible path (tangent lines) between two nodes, and return the
@@ -422,10 +284,7 @@ impl Pathfinder {
             //determine angle locations of sentinels
             let theta_s = ((r1.powi(2) + dist.powi(2) - r2.powi(2)) / (2f32 * r1 * dist)).acos();
             let phi_s = ((r2.powi(2) + dist.powi(2) - r1.powi(2)) / (2f32 * r2 * dist)).acos();
-            //println!(
-            //    "Generating Sentinels: Theta = {:?}, Phi = {:?}",
-            //    theta_s, phi_s
-            //);
+
             //sentinel vertices on A
             let a_s1 = theta_s;
             let a_s2 = -theta_s;
@@ -437,7 +296,6 @@ impl Pathfinder {
             let b_s3 = -PI + phi_s;
             let b_s4 = -PI - phi_s;
             sentinels = Some(vec![(a_s1, b_s1), (a_s2, b_s2), (a_s3, b_s3), (a_s4, b_s4)]);
-            //println!("{:?}", sentinels)
         }
 
         let mut connections = Vec::new();
@@ -501,176 +359,22 @@ impl Pathfinder {
         }
 
         // test for obstacles
+        let mut max_height = 0f32;
         for obstacle in &self.obstacles {
             // catch the simple cases for now: if a or b are inside the radius of obstacle, invalid
             // check if there are two points of intersect, for flyover cases
-            if let (Some(p1), Some(p2)) = self.perpendicular_intersect(a, b, obstacle)
-            {
-                // Intersect with obstacle, check if both start and end are above
-                println!("intersect with obstacle p1:{:?}, p2:{:?}", p1, p2);
-                if p1.z > obstacle.height && p2.z > obstacle.height {
-                    return PathValidity::Flyover(obstacle.height);
-                } else {
-                    return PathValidity::Invalid;
+            if let (Some(p1), Some(p2)) = perpendicular_intersect(&self.origin, a, b, obstacle) {
+                println!(
+                    "found intersection at height {} with obstacle {:?}",
+                    obstacle.height, obstacle
+                );
+                if obstacle.height > max_height {
+                    max_height = obstacle.height;
                 }
-                /*
-                // Minimum angle to fly over obstacles
-                let theta1 = match (a.z, b.z) {
-                    (ah, bh) if ah > bh => (p2.z - a.z).atan2(a.distance(&p2)),
-                    (ah, bh) if ah < bh => (p1.z - a.z).atan2(a.distance(&p1)),
-                    _ => 0f32,
-                };
-                if theta1 == 0f32 && a.z < obstacle.height {
-                    // If angle between waypoints is flat and less than height, cannot fly over
-                    return PathValidity::Invalid;
-                } else if theta_o < theta1 {
-                    // If angle between waypoints is less than required angle, cannot fly over
-                    return PathValidity::Invalid;
-                } else {
-                    // If angle between waypoints is greater than required angle, can fly over
-                    return PathValidity::Flyover(obstacle.height);
-                }
-                */
+                return PathValidity::Invalid; // Temporarily disable fly over
             }
         }
-        PathValidity::Valid
-    }
-
-    // temporary placeholder function to test functionality of point determination
-    pub fn perpendicular_intersect(
-        &self,
-        a: &Point,
-        b: &Point,
-        c: &Obstacle,
-    ) -> (Option<Point>, Option<Point>) {
-        // intersect distance gives x and y of intersect point, then distance
-        // calculates the shortest distance between the segment and obstacle. If less than radius, it intersects.
-        let (x, y, distance, endpoint) =
-            intersect_distance(a, b, &Point::from_location(&c.location, &self.origin));
-        if distance.sqrt() < c.radius as f32 {
-            // immediately check if the endpoint is the shortest distance; can't fly over in this case
-            // EXCEPTION: endpoint is inside obstacle but still generates a perpendicular.
-            // if endpoint {
-            //     // not technically none, but should be considered as such as we will stop calculations
-            //     return (None, None);
-            // }
-            let mag = (c.radius.powi(2) - distance).sqrt();
-            //println!("mag: {}", mag);
-            //calculate unit vectors for y and x directions
-            let dx = (a.x - b.x) / a.distance(b);
-            let dy = (a.y - b.y) / a.distance(b);
-
-            let p1 = Point::new(x + dx * mag, y + dy * mag, c.height);
-            let p2 = Point::new(x - dx * mag, y - dy * mag, c.height);
-            return (Some(p1), Some(p2));
-        } else {
-            return (None, None);
-        }
-    }
-
-    // Return intersection point(s) of line given by Point A and B and circle at point C with radius r
-    pub fn circular_intersect(
-        &self,
-        a: &Point,
-        b: &Point,
-        obstacle: &Obstacle,
-    ) -> (Option<Point>, Option<Point>) {
-        //y = mx + b for point a and b
-
-        let mut c = Point::from_location(&obstacle.location, &self.origin);
-        c.z = obstacle.height;
-        let dx = b.x - a.x;
-        let dy = b.y - a.y;
-
-        let (indep, dep, slope, slope_intercept) = if dx >= dy {
-            let indep = c.x;
-            let dep = c.y;
-            let slope = (b.y - a.y) / (b.x - a.x);
-            let slope_intercept = b.y - slope * b.x;
-            (indep, dep, slope, slope_intercept)
-        } else {
-            let indep = c.y;
-            let dep = c.x;
-            let slope = (b.x - a.x) / (b.y - a.y);
-            let slope_intercept = b.x - slope * b.y;
-            (indep, dep, slope, slope_intercept)
-        };
-
-        //Quadratic to solve for intersects
-        let quad_a = slope.powi(2) + 1.0;
-        let quad_b = 2.0 * (slope * slope_intercept - slope * dep - indep);
-        let quad_c = indep.powi(2) + dep.powi(2) + slope_intercept.powi(2)
-            - 2.0 * slope_intercept * dep
-            - obstacle.radius.powi(2);
-
-        //Check discriminant (if > 0, 2 intersects; if = 0, 1 intersect; if < 0, no intersects)
-        let discriminant = quad_b.powi(2) - 4.0 * quad_a * quad_c;
-
-        //Returning value of NAN for no solution points
-        if discriminant < 0.0 {
-            (None, None)
-        } else if discriminant == 0.0 {
-            let intersect_1: Point = if dx >= dy {
-                Point::new(
-                    (-1.0) * quad_b / (2.0 * quad_a),
-                    slope * ((-1.0) * quad_b / (2.0 * quad_a)) + slope_intercept,
-                    c.z,
-                ) //CURRENTLY JUST USES OBS HEIGHT
-            } else {
-                Point::new(
-                    slope * ((-1.0) * quad_b / (2.0 * quad_a)) + slope_intercept,
-                    (-1.0) * quad_b / (2.0 * quad_a),
-                    c.z,
-                ) //CURRENTLY JUST USES OBS HEIGHT
-            };
-            (Some(intersect_1), None)
-        } else
-        //if(discriminant > 0.0)
-        {
-            let (intersect_1, intersect_2) = if dx >= dy {
-                (
-                    Point::new(
-                        ((-1.0) * quad_b - (quad_b.powi(2) - 4.0 * quad_a * quad_c).sqrt())
-                            / (2.0 * quad_a),
-                        slope
-                            * (((-1.0) * quad_b - (quad_b.powi(2) - 4.0 * quad_a * quad_c).sqrt())
-                                / (2.0 * quad_a))
-                            + slope_intercept,
-                        c.z,
-                    ),
-                    Point::new(
-                        ((-1.0) * quad_b + (quad_b.powi(2) - 4.0 * quad_a * quad_c).sqrt())
-                            / (2.0 * quad_a),
-                        slope
-                            * (((-1.0) * quad_b + (quad_b.powi(2) - 4.0 * quad_a * quad_c).sqrt())
-                                / (2.0 * quad_a))
-                            + slope_intercept,
-                        c.z,
-                    ),
-                )
-            } else {
-                (
-                    Point::new(
-                        slope
-                            * (((-1.0) * quad_b - (quad_b.powi(2) - 4.0 * quad_a * quad_c).sqrt())
-                                / (2.0 * quad_a))
-                            + slope_intercept,
-                        ((-1.0) * quad_b - (quad_b.powi(2) - 4.0 * quad_a * quad_c).sqrt())
-                            / (2.0 * quad_a),
-                        c.z,
-                    ),
-                    Point::new(
-                        slope
-                            * (((-1.0) * quad_b + (quad_b.powi(2) - 4.0 * quad_a * quad_c).sqrt())
-                                / (2.0 * quad_a))
-                            + slope_intercept,
-                        ((-1.0) * quad_b + (quad_b.powi(2) - 4.0 * quad_a * quad_c).sqrt())
-                            / (2.0 * quad_a),
-                        c.z,
-                    ),
-                )
-            };
-            (Some(intersect_1), Some(intersect_2))
-        }
+        println!("path valid with threshold {}", max_height);
+        PathValidity::Flyover(max_height)
     }
 }
